VaultConfigurator: fixes

+++ src/interfaces/oracles/IRatiosOracle.sol
@@ -4,5 +4,5 @@ pragma solidity ^0.8.0;
 interface IRatiosOracle {
     function getTargetRatiosX96(
         address vault
-    ) external view returns (uint256[] memory ratiosX96);
+    ) external view returns (uint128[] memory ratiosX96);
 }
